+++ tests/test-partial-code-fence-imports.js
@@ -96,6 +96,7 @@ async function runTest() {
       undefined,
       true, // excludeImports
       false, // removeDuplicateHeadings
+      [], // preserveComponents
       undefined, // resolvedUrl
       undefined, // imageAssetMap
       undefined, // outDir